requirements.txt
streamlit
tensorflow==2.10.0
numpy
pillow
matplotlib
seaborn
scikit-learn
protobuf==3.20.3
